# Use safepoint to deliver SIGINT