--- a/src/cgutils.cpp
+++ b/src/cgutils.cpp
@@ -1717,3 +1717,20 @@ static Value *emit_exc_in_transit(jl_codectx_t *ctx)
     Constant *offset = ConstantInt::getSigned(T_int32, offsetof(jl_tls_states_t, exception_in_transit) / sizeof(void*));
     return builder.CreateGEP(pexc_in_transit, ArrayRef<Value*>(offset), "jl_exception_in_transit");
 }
+
+static void emit_signal_fence(void)
+{
+#ifdef LLVM39
+    builder.CreateFence(AtomicOrdering::SequentiallyConsistent, SingleThread);
+#else
+    builder.CreateFence(SequentiallyConsistent, SingleThread);
+#endif
+}
+
+static Value *emit_defer_signal(jl_codectx_t *ctx)
+{
+    Value *ptls = builder.CreateBitCast(ctx->ptlsStates,
+                                        PointerType::get(T_sigatomic, 0));
+    Constant *offset = ConstantInt::getSigned(T_int32, offsetof(jl_tls_states_t, defer_signal) / sizeof(sig_atomic_t));
+    return builder.CreateGEP(ptls, ArrayRef<Value*>(offset), "jl_defer_signal");
+}