--- a/src/codegen.cpp
+++ b/src/codegen.cpp
@@ -344,8 +344,8 @@ static GlobalVariable *jltls_states_func_ptr = NULL;
 static size_t jltls_states_func_idx = 0;
 #endif
 // Imaging mode only (non-imaging mode use pointer directly)
-static GlobalVariable *jl_gc_signal_page_ptr = NULL;
-static size_t jl_gc_signal_page_idx = 0;
+static GlobalVariable *jl_safepoint_page_ptr = NULL;
+static size_t jl_safepoint_page_idx = 0;
 
 // important functions
 static Function *jlnew_func;
@@ -3432,11 +3432,11 @@ static void allocate_gc_frame(BasicBlock *b0, jl_codectx_t *ctx)
     ctx->ptlsStates = builder.CreateCall(prepare_call(jltls_states_func));
     if (imaging_mode) {
         ctx->signalPage =
-            tbaa_decorate(tbaa_const, builder.CreateLoad(prepare_global(jl_gc_signal_page_ptr)));
+            tbaa_decorate(tbaa_const, builder.CreateLoad(prepare_global(jl_safepoint_page_ptr)));
     }
     else {
         ctx->signalPage = builder.CreateIntToPtr(
-            ConstantInt::get(T_size, (uintptr_t)jl_gc_signal_page), T_pint8);
+            ConstantInt::get(T_size, (uintptr_t)jl_safepoint_page), T_pint8);
     }
 }
 
@@ -5171,10 +5171,10 @@ static void init_julia_llvm_env(Module *m)
     }
 #endif
     if (imaging_mode) {
-        jl_gc_signal_page_ptr =
-            jl_emit_sysimg_slot(m, T_pint8, "jl_gc_signal_page.ptr",
-                                (uintptr_t)jl_gc_signal_page,
-                                jl_gc_signal_page_idx);
+        jl_safepoint_page_ptr =
+            jl_emit_sysimg_slot(m, T_pint8, "jl_safepoint_page.ptr",
+                                (uintptr_t)jl_safepoint_page,
+                                jl_safepoint_page_idx);
     }
 
     std::vector<Type*> args1(0);