--- a/src/debuginfo.cpp
+++ b/src/debuginfo.cpp
@@ -1674,9 +1674,7 @@ void RTDyldMemoryManagerUnix::registerEHFrames(uint8_t *Addr,
     di->start_ip = start_ip;
     di->end_ip = end_ip;
 
-    JL_SIGATOMIC_BEGIN();
     _U_dyn_register(di);
-    JL_SIGATOMIC_END();
 }
 
 void RTDyldMemoryManagerUnix::deregisterEHFrames(uint8_t *Addr,