This implements @vtjnash 's [idea](https://github.com/JuliaLang/julia/issues/14675#issuecomment-171810732) of delivering sigint in a safer way, without having to add sigatomic to everywhere. This also solves a few performance issues along the way. A brief summary of the differences,

* Make GC safepoint thread local

    And create 3 signal pages. See the comment in `safepoint.c` for detail.
    This have a very minor performance hit for accessing safepoint. However, I think we can avoid the safepoint and GC transition around gcframe setup and gcframe pop on x86 so this shouldn't be a big issue. (On ARM, the atomic release store is more expensive than two GC safepoint so we may use a normal store with two safepoint instead, the performance hit of making the safepoint thread local is still very minor in this case though (~5-10% for empty GC frame push/pop, cheaper than the tls getter call....)).

* Use the GC safepoint mechanism (i.e. SegFault) to deliver InterruptException at known points

    This automatically makes `ccall` of external library `sigatomic`. (Fix #1468, Fix #2622). In order to avoid waiting too long before the exception is delivered when running unmanaged code, this also implement waking up libuv, abort syscall and abort libsupport io (this use a hack but can be generalized if needed) when sigint arrives. This also implement force throwing of exception if SIGINT arrives too frequently so that one can use `Ctrl-C` to force abort some dead loop. A warning will be printed in such case since it bypass the safe path and even sigatomic.

    The most important consequence IMHO is that pressing `Ctrl-C` during sysimg compilation should no longer segfault =)

    This should also make it easier to implement #14675.

* Make parser and type inference calls sigatomic

    sigatomic is much cheaper (see below) now so it is now used to protect important runtime code to not be interrupted by `Ctrl-C`.

* Clean up safepoint and sigatomic, optimize try-catch and `sigatomic`

    * Split out `safepoint.c`, make `defer_signal` thread local and avoid the expensive atomic ops
    * Make `defer_signal` task local and automatically restore on expection. This eliminate the try-catch needed before in `disable_sigint`.

        This is technically breaking but `sigatomic_begin` and `sigatomic_end` aren't exported.

    * Inline `jl_sigatomic_begin` and `jl_sigatomic_end` in codegen.

    Benchmarking try-catch and `disable_sigint`

    ```jl
    julia> @inline f1() = try end
    f1 (generic function with 1 method)

    julia> @inline f2() = disable_sigint(()->nothing)
    f2 (generic function with 1 method)

    julia> g1(n) = for i in 1:n
               f1()
           end
    g1 (generic function with 1 method)

    julia> g2(n) = for i in 1:n
               f2()
           end
    g2 (generic function with 1 method)

    julia> function k(n)
               @time g1(n)
               @time g2(n)
           end
    k (generic function with 1 method)
    ```

    Timing before
    ```jl
    julia> k(10_000_000)
      0.264041 seconds
      0.401437 seconds
    ```

    Timing after
    ```jl
    julia> k(10_000_000)
      0.122592 seconds                                                                
      0.012506 seconds
    ```

Close https://github.com/JuliaLang/julia/pull/12333
Close #12309