@stevengj 2016-05-03
Suppose I have a `ccall` with a Julia callback function, e.g. `ccall(:my_numeric_integration, Cdouble, (Ptr{Void},), cfunction(my_julia_integrand, ...))`.    If I send a SIGINT while this is executing, does this mean that it will not interrupt the C code, but an `InterruptException` will still occur during the `my_julia_integrand` execution?

Also, if you have multiple `Task`s running, is there any way to indicate which `Task` you want to catch the exception?