@yuyichao 2016-05-03
A callback can still throw exception so sigatomic_begin/end are still useful. I made it fast because of this...

The sigint is always delivered to current task on the master thread if it doesn't block sigint (I realized that I might have missed a pending sigint check there, similar to the one in en_restore_state). However, you can block the sigint on the task you don't want it to be delivered by calling sigatomic_begin, is this good enough?