@stevengj 2016-05-03
I'm still confused by the callback situation in this PR.  If I *want* `InterruptExceptions` to be thrown during my callback (e.g. I have a `try-catch` block that handles them correctly), do I need to call `sigatomic_end/begin` in my callback?