@yuyichao 2016-05-03
For julia callback from C, I think a much faster sigatomic is as far as what this approach can provide.

For the task issue, I'm wondering if we can make non-root tasks sigatomic by default? This way any frontend will need to explicitly reenable sigint on tasks that want to handle sigint.