@yuyichao 2016-05-03
> (e.g. I have a try-catch block that handles them correctly), do I need to call sigatomic_end/begin in my callback?

A sigint can be delivered in the callback if the C caller wasn't called in sigatomic region. It is a little unreliably though since in principle a sigint can be delivered before the try block starts so I think to do this reliably you need to call the C function with sigatomic and reenable sigatomic in the callback after you started the try block.