@stevengj 2016-05-03
I'm not sure if non-root tasks should be sigatomic by default.  But it would be nice to have an option to make the task sigatomic when it is created.  e.g. `@async_nosigint`

I'm not sure how calling `sigatomic_begin/end` will help in Tasks, right now.  The problem is that `sigatomic_end` is still called within the Task, and as soon as it is called any deferred SIGINT will be thrown as an exception in that Task.  What if I don't want that Task to get `InterruptExceptions` at all?