@stevengj 2016-05-03
> you need to call the C function with sigatomic and reenable sigatomic in the callback after you started the try block.

Yes, that's what I'm doing now.