@vtjnash 2016-05-04
What is the sigatomic state of a new Task? e.g.
```julia
sigatomic_begin()
@async println("am I sigatomic?")
sigatomic_end()
yield()
```