@stevengj 2016-05-04
@yuyichao, `sigatomic_begin()` will only block SIGINT on the current task?  I thought it increments `jl_defer_signal`, which is global to all tasks?