@yuyichao 2016-05-05
Travis linux32 failure is https://github.com/JuliaLang/julia/issues/15920