@yuyichao 2016-05-05
Spelling mistakes fixed. Threading CI passed [AppVeyor](https://ci.appveyor.com/project/JuliaLang/julia/build/1.0.1224), [travis](https://travis-ci.org/JuliaLang/julia/builds/128043719). This should be ready for final review now.