@stevengj 2016-05-06
It seems like the documentation of the `sigatomic` functions should be revisited.  (I'd also like to see documentation of how to catch ctrl-c in callback functions, disable ctrl-c in certain tasks, etcetera.)