@yuyichao 2016-05-06
I've also noticed one case where `Ctrl-C` can still segfault. It is when we throw the error without any handler, and the execution of `on_exit` overflows the signal handler stack..... (at least on linux where we throw directly from the signal handler frame). A easy fix would be setting up a top level try catch frame at initialization time. The alternative would be to return to a call to `jl_throw` from the signal handler instead. However, it seems that `setcontext` or `makecontext` using a signal context might be undefined....