@yuyichao 2016-05-06
The only commit that touches libuv is the last one and reverting it seems to get rid of the problem for me locally (tested by piping it through `tee`) I'm checking now if what is the minimum change that triggers this now...