@yuyichao 2016-05-06
Something is fishy here. The buildbot seems to have been printing `jl_uv_writecb` warnings forever during compilation http://buildbot.e.ip.saba.us:8010/builders/package_win6.2-x64/builds/245/steps/make/logs/stdio The freezing happens on current master after printing the first warning. However, either with or without the commit reverted, the build passes without printing any warnings if `inference.ji` exists before I run make. (i.e. if I run make again after it freezes, the build passes without any warning).