@yuyichao 2016-05-07
> If you follow that issue

Which issue? I didn't see any mentioning of synchronization object in #11727?