@vtjnash 2016-05-07
you have to follow it somewhat far. there's a more concise explanation at https://github.com/libuv/libuv/issues/629, in which I demonstrate how it would make nodejs to hang (for the same reason) in a one-liner.