@yuyichao 2016-05-07
Ah, that one, I somehow think you are talking about `uv_async_t` since apparently simply initializing (not even using it) one is causing the trouble on windows......